@@ -1,7 +1,6 @@
 {
   "name": "gerritdrost/phenum",
   "description": "Because PHP needs Enums",
-  "minimum-stability": "dev",
   "license": "MIT",
   "authors": [
     {
@@ -10,11 +9,10 @@
     }
   ],
   "require": {
-    "php": ">=5.5.0"
+    "php": ">=8.1.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "4.5.0",
-    "satooshi/php-coveralls": "dev-master"
+    "phpunit/phpunit": "^10.0"
   },
   "autoload": {
     "psr-0": {
